Texture cube importer wraping up work
This CL add a way to specify if you want the cubemap (full image or
faces) being compressed (dds) when imported in cache, and if so you'd be
able to specify an image size if needed. (squared)

This also avoid re-importing / re-compressing everything when a cubemap
is found. Unless you specify the force Re-Import flag so you'd be able
to remove anything previously in cache.